+++ src/components/grouped-character/grouped-character.test.js
@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import GroupedCharacter from "./grouped-character.component";
 import CypressMountWithProviders from "../../../cypress/support/component-helper/cypress-mount";
 import {
@@ -179,6 +179,29 @@ context("Tests for GroupedCharacter component", () => {
           .and("have.css", "justify-content", `flex-${cssProp}`);
       }
     );
+  });
+
+  it.each(["10%", "30%", "50%", "80%", "100%"])(
+    "should check maxWidth as %s for GroupedCharacter component",
+    (maxWidth) => {
+      CypressMountWithProviders(
+        <GroupedCharacterComponent maxWidth={maxWidth} />
+      );
+
+      getDataElementByValue("input")
+        .parent()
+        .parent()
+        .should("have.css", "max-width", maxWidth);
+    }
+  );
+
+  it("when maxWidth has no value it should render as 100%", () => {
+    CypressMountWithProviders(<GroupedCharacterComponent maxWidth="" />);
+
+    getDataElementByValue("input")
+      .parent()
+      .parent()
+      .should("have.css", "max-width", "100%");
   });
 
   describe("check events for GroupedCharacter component", () => {