+++ src/components/multi-action-button/multi-action-button.test.js
@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import MultiActionButton from "./multi-action-button.component";
 import Button from "../button";
 import { Accordion } from "../accordion";
@@ -17,6 +17,7 @@ import {
   CHARACTERS,
 } from "../../../cypress/support/component-helper/constants";
 import CypressMountWithProviders from "../../../cypress/support/component-helper/cypress-mount";
+import { useJQueryCssValueAndAssert } from "../../../cypress/support/component-helper/common-steps";
 
 const testData = [CHARACTERS.DIACRITICS, CHARACTERS.SPECIALCHARACTERS];
 const MultiActionButtonList = ({ ...props }) => {
@@ -78,15 +79,15 @@ context("Tests for MultiActionButton component", () => {
     );
 
     it.each([
-      [SIZE.SMALL, "32px"],
-      [SIZE.MEDIUM, "40px"],
-      [SIZE.LARGE, "48px"],
+      [SIZE.SMALL, 32],
+      [SIZE.MEDIUM, 40],
+      [SIZE.LARGE, 48],
     ])("should render Multi Action Button with %s size", (size, height) => {
       CypressMountWithProviders(<MultiActionButtonList size={size} />);
 
-      multiActionButtonComponent()
-        .should("have.css", "height")
-        .and("contain", height);
+      multiActionButtonComponent().then(($el) => {
+        useJQueryCssValueAndAssert($el, "height", height);
+      });
     });
 
     it.each(["left", "right"])(
@@ -134,7 +135,14 @@ context("Tests for MultiActionButton component", () => {
       pressTABKey(1);
       multiActionButtonComponent()
         .children()
-        .should("have.css", "border", "3px solid rgb(255, 181, 0)");
+        .then(($el) => {
+          const values = $el.css("border").split(" ");
+          expect(parseInt(values[0])).to.be.within(1, 3);
+          expect(values[1]).to.equals("solid");
+          expect(`${values[2]}${values[3]}${values[4]}`.trim()).to.equals(
+            "rgb(255,181,0)"
+          );
+        });
     });
 
     it("should render Multi Action Button with specific background colour when hovering", () => {