+++ src/components/pager/pager.style.js
@@ -1,6 +1,7 @@
 import styled, { css } from "styled-components";
 import StyledInput from "../../__internal__/input/input.style";
 import StyledInputPresentation from "../../__internal__/input/input-presentation.style";
+import StyledFormField from "../../__internal__/form-field/form-field.style";
 import InputIconToggleStyle from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import StyledSelectText from "../select/__internal__/select-text/select-text.style";
 import Link from "../link";
@@ -95,6 +96,10 @@ const StyledPagerNavInner = styled.div`
   display: flex;
   align-items: center;
   padding: 0 12px;
+
+  && ${StyledFormField} {
+    margin-bottom: 0;
+  }
 `;
 
 const StyledPagerLinkStyles = styled(Link)`