+++ src/components/search/search.test.js
@@ -105,8 +105,8 @@ context("Test for Search component", () => {
     );
 
     it.each([
-      ["34%", "458px"],
-      ["70%", "944px"],
+      ["34%", "464px"],
+      ["70%", "956px"],
     ])(
       "should render Search with searchWidth prop set to %s",
       (widthInPercentage, widthVal) => {
@@ -116,26 +116,58 @@ context("Test for Search component", () => {
 
         searchDefault().then(($el) => {
           expect($el[0].getBoundingClientRect().width).to.be.within(
-            parseToIntElement(widthVal),
+            parseToIntElement(widthVal) - 1,
             parseToIntElement(widthVal) + 2
           );
         });
       }
     );
 
-    it.each([["475px"], ["250px"]])(
+    it.each(["475px", "250px"])(
       "should render Search with searchWidth prop set to %s",
       (width) => {
         CypressMountWithProviders(<SearchComponent searchWidth={width} />);
 
         searchDefault().then(($el) => {
           expect($el[0].getBoundingClientRect().width).to.be.within(
-            parseToIntElement(width),
+            parseToIntElement(width) - 1,
             parseToIntElement(width) + 2
           );
         });
       }
     );
+
+    it.each([
+      ["10%", "135px"],
+      ["34%", "464px"],
+      ["70%", "956px"],
+      ["100%", "1366px"],
+    ])(
+      "should render Search with maxWidth prop set to %s",
+      (widthInPercentage, widthVal) => {
+        CypressMountWithProviders(
+          <SearchComponent maxWidth={widthInPercentage} />
+        );
+
+        searchDefault().then(($el) => {
+          expect($el[0].getBoundingClientRect().width).to.be.within(
+            parseToIntElement(widthVal) - 1,
+            parseToIntElement(widthVal) + 2
+          );
+        });
+      }
+    );
+
+    it("when maxWidth has no value it should render as 100%", () => {
+      CypressMountWithProviders(<SearchComponent maxWidth="" />);
+
+      searchDefault().then(($el) => {
+        expect($el[0].getBoundingClientRect().width).to.be.within(
+          parseToIntElement("1366") - 1,
+          parseToIntElement("1366") + 2
+        );
+      });
+    });
 
     it.each([
       ["default", "rgb(102, 132, 148)"],